+++ apps/app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.6](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.18.5...vtj-project-app@0.18.6) (2026-06-22)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
+## [0.18.5](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.18.4...vtj-project-app@0.18.5) (2026-06-22)
+
+**Note:** Version bump only for package vtj-project-app
+
+
+
+
+
 ## [0.18.4](https://github.com/samchen08/vtj.pro/compare/vtj-project-app@0.18.3...vtj-project-app@0.18.4) (2026-06-22)
 
 **Note:** Version bump only for package vtj-project-app