+++ platforms/pro/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.18.6](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.18.5...@vtj/pro@0.18.6) (2026-06-22)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
+## [0.18.5](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.18.4...@vtj/pro@0.18.5) (2026-06-22)
+
+**Note:** Version bump only for package @vtj/pro
+
+
+
+
+
 ## [0.18.4](https://gitee.com/newgateway/vtj/compare/@vtj/pro@0.18.3...@vtj/pro@0.18.4) (2026-06-22)
 
 **Note:** Version bump only for package @vtj/pro